pyproject.toml
[tool.poetry]
name = "dlabs_ai_test"
version = "0.1.0"
description = ""
authors = ["Zbigniew <zbigniew.polito@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.7.3"
fastapi = "^0.63.0"
uvicorn = "^0.13.4"
python-multipart = "^0.0.5"
aiofiles = "^0.6.0"
aiosqlite = "^0.17.0"
Pillow = "^8.1.0"
aiocache = "^0.11.1"
ujson = "^4.0.2"
asynctempfile = "^0.5.0"
boto3 = "^1.17.19"
awscli = "^1.19.19"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
create = true
in-project = true
